pyproject.toml
[project]
name = "ai-data-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "langchain-core>=1.2.7",
    "langsmith>=0.6.7",
    "numpy>=2.4.1",
    "openai>=2.16.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
